fix https://github.com/nervosnetwork/ckb-miscellaneous-scripts/pull/3
Except: ckb-c-stdlib support declaration and definition splitting. It will be in another pull request.
